refactor: move passthroughview to a separated file
