+++ Sources/Clappr_iOS/Classes/Plugin/Core/DoubleTap/DoubleTapPlugin.swift
@@ -36,16 +36,12 @@ public class DoubleTapPlugin: UICorePlugin {
     private func bindCoreEvents() {
         guard let core = core else { return }
         listenTo(core, eventName: Event.didChangeActiveContainer.rawValue) { [weak self] _ in self?.bindEvents() }
-        listenTo(core, eventName: Event.didShowModal.rawValue) { [weak self] _ in self?.didShowModal() }
-        listenTo(core, eventName: Event.didHideModal.rawValue) { [weak self] _ in self?.didHideModal() }
-    }
-    
-    private func didShowModal() {
-        doubleTapView.isUserInteractionEnabled = false
-    }
-    
-    private func didHideModal() {
-        doubleTapView.isUserInteractionEnabled = true
+        listenTo(core, eventName: Event.didShowModal.rawValue) { [weak self] _ in
+            self?.doubleTapView.isUserInteractionEnabled = false
+        }
+        listenTo(core, eventName: Event.didHideModal.rawValue) { [weak self] _ in
+            self?.doubleTapView.isUserInteractionEnabled = true
+        }
     }
     
     override public func render() {
@@ -105,41 +101,5 @@ public class DoubleTapPlugin: UICorePlugin {
         if #available(iOS 10.0, *) {
             UIImpactFeedbackGenerator().impactOccurred()
         }
-    }
-}
-
-class PassthroughView: UIView {
-    
-    var core: Core?
-    
-    override func point(inside point: CGPoint, with event: UIEvent?) -> Bool {
-        if isPosterPluginVisible() {
-            return false
-        }
-        return mediaControlPluginsColidesWithTouch(point: point, event: event)
-    }
-    
-    private func isPosterPluginVisible() -> Bool {
-        let posterPlugin = core?.activeContainer?.plugins.first(where: { $0 is PosterPlugin })
-        return !(posterPlugin?.view.isHidden ?? false)
-    }
-    
-    private func mediaControlPluginsColidesWithTouch(point: CGPoint, event: UIEvent?) -> Bool {
-        guard let mediaControlPlugin = core?.plugins.first(where: { $0 is MediaControl }) else { return false }
-        
-        let pluginColidingWithGesture = filteredOutModalPlugins()?.first(where: {
-            pluginColideWithTouch($0, point: point, event: event)
-        })
-        
-        return pluginColidingWithGesture == nil || mediaControlPlugin.view.isHidden
-    }
-    
-    private func filteredOutModalPlugins() -> [UICorePlugin]? {
-        let pluginsWithoutMediaControl = core?.plugins.filter({ $0.pluginName != MediaControl.name })
-        return pluginsWithoutMediaControl?.filter({ ($0 as? MediaControlPlugin)?.panel != .modal })
-    }
-    
-    private func pluginColideWithTouch(_ plugin: UICorePlugin, point: CGPoint, event: UIEvent?) -> Bool {
-        return plugin.view.point(inside: self.convert(point, to: plugin.view), with: event)
     }
 }
